requirements.txt
flask
flask-cors
# New for auth testing
flask-login
flask-sqlalchemy
flask-bcrypt
# Probably switch this to using flask-MySQLdb and MySQLdb
# To install mariadb: https://askubuntu.com/questions/1438002/mariadb-connector-c-is-not-installed
# sudo apt install mariadb-client libmariadb3 libmariadb-dev
# Flask-SQLAlchemy

# Encryption stuff
# cryptography
# rsa
# pycryptodome

# Disabled db for testing

# PyMySQL
# PyOpenSSL

# mariadb
# This doesn't seem to be needed on linux.
# secrets
# gnupg
# Html
beautifulsoup4
htmlmin
validators

python-dotenv

# WGSI Server
waitress

requests
